temporarily rewrite the script in zsh to fix a critical bug
The bug was that ERR_EXIT wasn't affecting a subshell in which repos are
downloaded. I wrapped that part like that to have a single error
handler, but apparently it wasn't even being invoked. In any case, the
git clone command would fail on non-existent repos, but the execution
would continue, after that the tar command would be invoked creating an
empty archive because tar writes an empty file even if the inputs don't
exist (in other words, it doesn't check inputs preemptively and
immediately begins streaming the archive to disk). As such, I've lost a
few repos, most of which I've been able to recover safely.

Zsh doesn't have this problem (I hope), but only if I use functions,
with its ERR_RETURN option.